backend/requirements.txt
langchain
langchain-classic
langchain-openai
langchain-community
langchain-huggingface
langchain-text-splitters
sentence-transformers
chromadb
fastapi
uvicorn[standard]
pypdf
python-dotenv
# qdrant-client
python-multipart